requirements.txt
flask
flask_restful
firebase-admin
Pillow
numpy
requests